make sleep work despite suspend/resume
When you say "sleep 3600", then put the system into suspend state, the sleep
counter stops ticking -- the code only wakes up an hour after you resume the
system.

We want this to be more "1 hour from now" than "1 hour of actual sleep".

(Note that we still have an upto-60-seconds overshoot, but we deem that to be
ok).